backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
alembic==1.13.3
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
pydantic==2.9.2
pydantic-settings==2.5.2
apscheduler==3.10.4
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
playwright==1.47.0
tenacity==9.0.0
